test(moderation): fold verification tests into the adminApi suite
The new cases covered adminApi.ts but landed in a second file, next to an
adminApi.test.ts that already holds a verifyAgeRestricted block, so tests
for one module were split across two files with two mocking setups.

Moving them also lets them use that file's established fake-timer pattern
instead of sleeping through the real 500ms/1000ms/5000ms waits: the ten
cases drop from 12.5s to under a millisecond, with the timeout path alone
having accounted for 6s. All five fail-closed assertions still go red
against the pre-fix adminApi.ts, so the move preserved discrimination.
